+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-server: log process status updates;
 - concord-server: initial support for Teams;
 - concord-server: an endpoint to retrieve a list of attachments.